+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### v0.0.53
 
+- Notification Center - Add new component
+- Switch - Add new component
 - Card List Item - Added prop `eventDate`
 - Card List Item - Added prop `eventTime` 
 - Card List Item - Added prop `timePeriod`